workflow: build release image
Automatically trigger an image build whenever a new version is released.
The 'latest' tag is conditional and depends on whether the "Set as the
latest release" option is checked or not in Github.

This introduces a small behavior change: there is now a 1:1 relationship
between the tag in git and for the container; we used to remove the 'v'
before.

Signed-off-by: Antoine Tenart <atenart@redhat.com>