retis-events: remove non-existent pyo3 feature dep
'pyo3' is listed as an extra feature dependency to the 'python' one.
However retis-event does not have a 'pyo3' feature. It's not a way to
select the 'pyo3' crate either as this is done in the 'python-embed' and
'python-lib' features. The 'python' feature is just an underlying one
for shared behavior between the other more specialize Python features.

Remove it.

Signed-off-by: Antoine Tenart <atenart@redhat.com>