events: formatting: fix newline at end of buffer handling
When a newline is found at the end of the buffer it was printed twice.
This is because splitting such string will result in having an empty
last split. The logic was then printing a new line for the split and
because one was at the end of the buffer.

This also prevents from printing the prefix and fixes cases where a
double newline is found in the middle of the buffer.

Fixes: d5350df26224 ("events: add our own formatter")
Reported-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: Antoine Tenart <atenart@redhat.com>