ovs: flow_info: implement enrichment logic
The main concern here is to balance a between a reasonable enrichment
rate and avoid overloading OVS.

Calls to OVS are throttled and if f events come at a higher pace than
what the enricher thread can handle, old requests are dropped.

In addition, a registry keeps track of the UFIDs that have been
enriched, to avoid doing so twice. For this registry to work sf_acts and
flow pointers are added. This is because a UFID just represents the
key of the flow (the match part). By adding the sf_acts and flow
pointers we make sure the full flow is still the same.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>