ovs: add support for drop action
Drop action has been added recently. This means that we could find a
kernel that does not support it and still supports running retis.

On such kernels, we need to determine if an action type value of 24
(currently used for drop) might mean SET_TO_MASKED. In order to do that,
create a global config array and place the value of
__OVS_ACTION_ATTR_MAX. Recent actions should only be emitted if their
value is less than __OVS_ACTION_ATTR_MAX.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>