ovs: add support for drop action
Drop action has been added recently. This means that we could find a
kernel that does not support it and still supports running retis.

For that reason, the value to check against must be dynammic, so
changing the switch/case into an if/else one.

This patch includes functional tests that are skipped if kernel or OVS
do not support the feature.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>