events: better Python representation of complex enums
Data inside complex enums can't be directly accessed in the Python
representation of the events. The only way was to use the dict
representation, which prevents from using the event fields directly.
This made the enum fields to stand out, including the whole OvS event.

Instead do a little dance with serde and allow in Python things like:

  print(e['skb'].ip.version.v6.flow_label)

This does not change the JSON representation.

Closes #437.

Suggested-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: Antoine Tenart <atenart@redhat.com>