core: probes: fix segfault on exit
A segmentation fault can happen when exiting a collection and using USDT
probes. This is because a manager is allocated in libbpf when using USDT
probes, which must lives until the probe link is dropped as it is then
freed.

The segfault resurfaced as the mentioned commit actually reverted a
previous commit fixing the same issue, 9baa6fc2ca7f ("usdt: remove
ManuallyDrop<> workaround.").

Fix this by keeping the skel alive during the collection. This did not
occur for raw tracepoints as there is no manager in libbpf for non-USDT
probes, but for completeness also do the change there.

Fixes #475.

Fixes: 125136798f5f ("tree: update libbpf-rs/sys")
Reported-by: Paolo Valerio <pvalerio@redhat.com>
Signed-off-by: Antoine Tenart <atenart@redhat.com>