collector: skb: handle VLAN in the payload
Collecting packets with one or more VLAN header(s) in their payload
resulted in packets we could not parse. Fix this by skipping the VLAN
headers if we found them. The VLAN information is reported in the event
but is coming from the BPF part directly (to support h/w acceleration).

Signed-off-by: Antoine Tenart <atenart@redhat.com>