ct: include ct->status in the events
This is very useful and must be part of the event as status is a
crucial part of the nf_conn representation.

The field is a bitmask, a potentially fairly big one, so pretty
printing it might lead to verbose output even for basic cases and so
it's kept as hex. This is safe as those definitions are uapi and do
not require and so stable across kernel versions.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>