collect/process: add a flag to print LL information
Previous logic relied on whether or not a LL sub-section is present. But
we know always have this information in the raw packet and we're moving
away from parsing the packet ourselves. This is better as it allows to
decide if the LL information is important at post-processing time.

The one drawback is the Python `show()` helper which isn't configurable
yet in this regards (it always prints LL info).

Signed-off-by: Antoine Tenart <atenart@redhat.com>