collect/process: add a flag to print LL information
Stop relying on event sections being present for displaying LL information. Use
a cli option instead. While doing so we also make the VLAN section to always be
reported, as we can't print the Ethernet information without the VLAN one when
the LL reporting option is set. This will also be a better setting with the
upcoming patches as VLAN info can come from both the packet and the metadata,
and we don't want to only have one of the two.

The one drawback is the Python `show()` helper which isn't configurable yet in
this regards (it always prints LL info).

Signed-off-by: Antoine Tenart <atenart@redhat.com>