cli: move kconf option to subcommands
Early initializing of the inspector based on kconf option is only used
in "collect" and "inspect". Move the option there (no need to keep it
global), and remove the ugly hack in main.rs.

In order to keep `retis_in_container.sh` working, remove the use of
RETIS_KCONF and instead map all well known config paths and suggest
users to use the `--kconf` option as failback.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>