tree: provide the main config to runners
Recent changes made it impossible to check the main configuration
options from runners, which made things a lot cleaner. Also we moved to
providing the full command options to collectors, which is not 100%
related to this but shows the direction: letting all parts access all
the cli options.

This provides the main config to runner, so those options can be checked
from the runners internal logic.

Signed-off-by: Antoine Tenart <atenart@redhat.com>