collect: configure section_factories during init
This is a continuation of discussions that took place during [1].

Some collectors' configuration might affect the way section factories
operate. Pass the section factories to collectors during init() phase so
they can pass such configuration.

In order to make use of factories easier, make it a proper type.

The skb collector is the first user of this functionality.

[1] https://github.com/retis-org/retis/pull/464

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>