core: probe: builder: attach all probes at once
Instead of attaching all probes one by one, add them to the builders
(transferring ownership, which is more logical) and only then attach all
probes at once. Attaching probes added to a builder does not consume the
builder, it can be reused for later probe(s).

This does not change the current behavior but opens the door for
optimizations.

Signed-off-by: Antoine Tenart <atenart@redhat.com>