core: probe: kernel: set the symbol address in the BPF cookie
The kernel symbol address is used as the primary probe configuration
key. It could be retrieved dynamically in k(ret)probes but not in
raw_tracepoints and was there set as rodata. In an effort to move away
from probe-specific rodata, set the symbol address as the BPF cookie
when attaching all probes and use it to get the symbol address.

When setting cookies is not supported, fallback to the old behavior.
This is mainly for c8s.

Signed-off-by: Antoine Tenart <atenart@redhat.com>