core: probe: kernel: tp: reuse the same program for tp sharing the same nargs
Raw tracepoints are special as the verifier checks at load time no out
of bound argument is accessed, which prevents us from loading a program
once and reusing it for all raw tracepoint probes. However since we
moved the symbol address to the BPF cookie the only probe-specific
variable in our tracepoints is the number of arguments. This one has to
be rodata for the verifier to work, but nothing prevents us from sharing
the same program for tracepoints have the same number of arguments.

Do this. This speed up loading time a lot when using tracepoints, e.g.
the generic profile startup time is divided by two as we have the
following and as loading a raw tracepoint program always take quite some
time:

 nargs 1: [
        "tp:net:napi_gro_receive_entry",
        "tp:net:napi_gro_frags_entry",
        "tp:net:netif_rx",
        "tp:net:net_dev_queue",
        "tp:net:netif_receive_skb"]
 nargs 2: ["tp:net:net_dev_start_xmit"]
 nargs 4: ["tp:skb:kfree_skb"]

Signed-off-by: Antoine Tenart <atenart@redhat.com>