events: make EventSectionFactory Send + Sync
This will ensure we can use the factories in different threads.

We kept `as_any_mut` as this is only used during initialization where factories
are not yet shared in threads. We could move to using internal mutability and
locks, at the expense of taking a read lock in the fast path later.

Signed-off-by: Antoine Tenart <atenart@redhat.com>