core: events: use a bounded channel for events
Not using a bounded buffer is a major issue as events can keep growing even if
they have no chance of being handled. This makes memory consumption to increase,
newer events to be delayed and to not reflect the reality of what is being
captured.

We should instead always use a channel with a known maximum capacity to be
explicit about what we can handle and to allow pushing back on the source of the
data (eBPF collection).

Signed-off-by: Antoine Tenart <atenart@redhat.com>