core: probe: kernel: fix verifier error
For some reason code compiled with clang 20 (e.g. on Rawhide) results in
the BPF verifier thinking an oob access is done when accessing arguments
from our context. This is a false positive that is fixed by inverting
the checks in retis_get_param.

This is a zstream 1.5 fix only as code in main is not impacted.

Signed-off-by: Antoine Tenart <atenart@redhat.com>