ovs: generate flow lookup miss events
If the flow lookup results in a flow miss (flow = 0), we currently do
not generate an OVS event. This is confusing as we see a kernel event
without OVS section. Also, we don't send the stats which might be
useful.

This patch makes the probe emit an event with an empty flow that will
not get enriched but will be displayed as "flow miss".

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>