Makefile: do not remap SELinux labels
Remapping files live is not best practice (this can cause issues, has
poor performances, etc). When using a container the files must be mapped
correctly before.

However here this is not practical as the source files are shared
between different systems (host, container). Also the container is only
here to get the right set of tools. In such cases the recommendation is
simply to disable the use of labels.

Signed-off-by: Antoine Tenart <atenart@redhat.com>