tests: convert raw packet to JSON
Now that we don't parse the raw packet and expose some fields in the event, the
CI can't check the packet's content. In order to do this we now have to parse
the raw packet using Scapy at post-processing time. The logic here parses raw
packets in the events and converts the Scapy output into a JSON representation
for later field checking.

Signed-off-by: Antoine Tenart <atenart@redhat.com>