profiles: allow providing a path to a profile
In order to be used profiles had to be stored in known paths. This is
fine for lasting profiles or for distributing them but there are cases
where it's not convenient:

- When developing a new profile.
- When sharing a temporary profile (e.g. while debugging an issue with
  someone).

In order to better support those cases make the profile option to also
support providing a full path to a (single-)profile file.

Signed-off-by: Antoine Tenart <atenart@redhat.com>