collect: remove auto-mode
At first auto-mode was how we referred to the handling of collectors if
not explicitly enabled: try using if prerequisites are met. Over time
the concept was extended to default probes and possible future
additions. It was still not well defined as even between the two only
users the set of checks was not the same.

It turns out this led to incorrect and unexpected behavior:

- `retis -p <profile> collect` would fail if some prerequisites weren't
  available (e.g. OvS) as it relied on "auto-mode" for collectors.

- "Auto-mode" for collectors did not include the probes being set, while
  "auto-mode" for probes was depending on the "collectors" behavior.

Taking a step back, it seems "auto-mode" is ill-defined and can't
actually be shared between different parameters. Even now it's not. What
we're trying to achieve is actually defining the default behavior of
parameters and making some of them smart. This is fine and does not
require a global "auto-mode" flag.

This allows to use a profile while relying on the default behavior for
collectors, which is quite handy. On the other hand probes must be
explicitly defined if a profile is used, to allow using only the probes
set by the collectors if wanted and to improve their expected behavior.

For completeness: "auto-mode" could make sense if shared by different
parameters, and is making sense for the default set of probes. The
collector behavior has nothing to do with it and is only the default
parameter behavior. But since there would be a single user of
"auto-mode" and since we don't know if it would actually make sense one
day; let's just remove a concept that has no justification.

Signed-off-by: Antoine Tenart <atenart@redhat.com>