Makefile: strictly control CFLAGS for building BPF programs
Sharing the same CFLAGS values for the Rust and BPF parts is convenient
but cannot be reliable. Many flags are incompatible with the BPF target
which causes issues when building Retis in build environments (e.g.
distributions or build systems) setting common CFLAGS for across
packages.

Fix this by strictly controlling how we build the BPF part, ignoring
CFLAGS passed by the caller.

Signed-off-by: Antoine Tenart <atenart@redhat.com>