Makefile: allow unused cmdline args for BPF compilation
When compiling in build system environments, e.g. Buildroot, wrappers
might be used to call compilers. In such cases it's not possible to
control the exact list of arguments and issuing errors on unused command
line arguments will make the build to fail.

As unused command line arguments are harmless, let's not issue errors.

Signed-off-by: Antoine Tenart <atenart@redhat.com>