Makefile: do not ship the asm header workaround
In the Makefile we add the /usr/include/x86_64-linux-gnu path to the
list of include folders by default. This is to fix a known particularity
of Ubuntu where the asm folder is not set up as expected by compilers.

"""
In file included from usdt.bpf.c:3:
In file included from /tmp/cirrus-ci-build/retis/src/.out/bpf/usdt.bpf.h:6:
/usr/include/linux/errno.h:1:10: fatal error: 'asm/errno.h' file not found
         ^~~~~~~~~~~~~
1 error generated.
"""

This should instead be fixed in Ubuntu, or manually worked around by
users. Remove the workaround from Retis and add it back for the CI parts
running on Ubuntu. While at it use a more generic folder.

Signed-off-by: Antoine Tenart <atenart@redhat.com>