events: remove EventSection trait
There is no need to treat EventSections as a common trait. For python
handling, we have pyo3::Py and for json-related stuff we have
serde::Serialize.

Since Event is Serialize, Series is as well. So converting them to/from
json can be done directly on the types without going through
serde_json::Value.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>