pcap: adjust the capture time adding the initial offset event
All the timestamps are not fixed up by applying the initial offset
making, in some cases, it extremely annoying to correlate events with
the pcapng files generated.
That aside, this way the actual capture time is shown in the packets
other than the offset from boot (epoch in the pcapng case).

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>