collector: skb: retrieve the netns cookie
Network namespaces are identified by their inum and those values are
useful to link events to actual network namespaces (e.g. containers);
but they have a major issue: they are not guaranteed to be unique and
can be reused after netns removal.

In addition to the inum, retrieve the unique identifier (cookie) and
report it in the events. This can be used to distinguish between netns
in a reliable way.

While at it rename the field storing the inum.

Signed-off-by: Antoine Tenart <atenart@redhat.com>