collector: skb: always retrieve dev and ns data
There are not many cases where one would not want to get dev and netns
information, as the vast majority of deployments are using containers
and/or many networking interfaces (PF, VF, multiple cards, etc).

Let's just always report dev and netns information. The overhead should
be small. Note that dev was already the default value of --skb-sections.

Signed-off-by: Antoine Tenart <atenart@redhat.com>