tree: use tracefs instead of debugfs when possible
Initially tracefs was part of debugfs and only accessible when debugfs
was mounted. However this changed and tracefs can now be mounted even if
debugfs is not. This is now a common behavior by default in many
distributions.

This is indeed better; in Retis try using the standalone tracefs mount
instead of the nested one.

Signed-off-by: Antoine Tenart <atenart@redhat.com>