tools: rewrite the gen_vmlinux script
Generating a unified vmlinux.h file from BTF is tricky when more than a
one module is needed. As the module BTF are split BTFs, conflicts in
definitions and re-definitions of types can happen, which is hard to
detect and remove. We used a simple logic, using diff, but that turned
out not to work as expected.

Instead define a list of types we use in Retis and use pahole to
generate the minimum set of definitions required for compilation. This
can be used using multiple modules and avoid conflicts and
re-definitions.

The drawback is having to maintain a list of types we're interested of,
but that is probably the least bad approach.

Signed-off-by: Antoine Tenart <atenart@redhat.com>