core: tracking: fix spurious events being reported
When probing the return of a function (kretprobe), the skb is processed
as usual and hits the filtering + tracking logic. This is an issue if it
was freed in the function, as a new tracking id would be inserted (for
an already freed event) and matched later for unrelated packets if the
memory is reused.

Fix this by disallowing inserting new tracking information from
kretprobes. Getting an existing tracking id is not impacted.

This can lead in rare cases to an event from such probe not being linked
to later events, if the skb wasn't freed and was seen for the first
time.

Reported-by: Paolo Valerio <pvalerio@redhat.com>
Signed-off-by: Antoine Tenart <atenart@redhat.com>