pcap: use probe info as interface
The device and namespace information is already present in the capture
metadata.
Besides, without probe information, seeing multiple identical packets on
the same interface might be interpreted as a retransmission. For retis,
an interface (i.e: a point where packets are captured) is equivalent to
the probe name. Use that instead of the device if no probe filter has
been provided.

TODO: add tests

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>