events: move packet to its own top-level section
The raw packet event is moved into its own top level section. This makes sense
as the packet itself isn't necessarily linked to an skb.

For now we're still using the per-collector factories and converting a
collector-specific BPF event to the generic PacketEvent. This is only a few
additional lines and might help supporting different cases in the future, as the
important part is really about the representation in the Retis event. This can
easily be revisited later.

Signed-off-by: Antoine Tenart <atenart@redhat.com>