collect: move the startup event generation to the file writer
The startup event is only used at post-processing time but is generated even of
events are not stored to the disk. Moreover the output file can be rotated and
in such case a new startup event should be issued. To improve the experience in
both cases, transfer the startup event generation to the file writer
(RotateWriter).

Signed-off-by: Antoine Tenart <atenart@redhat.com>