events: allow to check if section can print anything
With the move of the packet section to the top level event, the skb section
might not print anything in case LL information isn't requested. This is however
a common problem and depending on the sections and what they contain, we can end
up with sections not printing anything. To help with that add an helper to allow
them to report if anything would be printed.

Signed-off-by: Antoine Tenart <atenart@redhat.com>