events: python: rename the raw method to to_raw
This is to avoid conflict with the packet.raw attribute. Naming it with the
"to_" prefix helps transmitting the intent and what happens underneath
(conversion to a dict representation).

While at it remove outdated Python examples from the Rust doc, to help
maintenance (documentation should stay but we already have examples in the
official doc).

Signed-off-by: Antoine Tenart <atenart@redhat.com>