events: python: rename the raw method to to_dict
Naming it with the "to_" prefix helps transmitting the intent and what happens
underneath (conversion to a dict representation) and removing the use of "raw"
makes things more clear.

While at it remove outdated Python examples from the Rust doc, to help
maintenance (documentation should stay but we already have examples in the
official doc).

Signed-off-by: Antoine Tenart <atenart@redhat.com>