events: python: unify the top level event and the sections implementation
Helpers such as __repr__ were defined in a generic way for event sections but
this was not shared with the top level event which had its own implementation
(in its wrapper, PyEvent). This led to having a different behavior for the
__repr__ helper for example. It also makes things harder to maintain.

The reason behind that was how the event was defined previously which prevented
a generic implementation. But this is not longer the case as it is now a normal
structure. This means we can now derive `event_section` on the top level event
directly and automatically get the same generated helpers.

Once the above is done we can get rid of the specific implementation of the
PyEvent helpers and just forward the ones we can to the inner event.

While at it define to 'to_raw()' helper for event types and not only sections.
The rework has the side effect of fixing the __repr__ implementation for event
sections.

Signed-off-by: Antoine Tenart <atenart@redhat.com>