events: increase max size in the BPF map
When enabling all event sections the reserved size is not wide enough in the
map. Increase it; the events map now is 16MB. This is fine for now but we might
want to make this a bit more dynamic in the future.

At the same time export event defines from BPF to Rust for easier maintenance.

Signed-off-by: Antoine Tenart <atenart@redhat.com>