probe: raw_tp: fix cookie support detection
Cookie support in raw tracepoints was done looking at if the
bpf_get_attach_cookie_tracing type existed, but this[1] actually precede the
introduction of the raw tp cookie support. Fix this by also checking if the
cookie field exist in the bpf_raw_tp_link struct instead, which was introduced
as part of the raw tp cookie support[2].

Fixes #560.

[1] 2fcc82411e74 ("bpf, x86: Attach a cookie to fentry/fexit/fmod_ret/lsm.").
[2] 68ca5d4eebb8 ("bpf: support BPF cookie in raw tracepoint (raw_tp, tp_btf)
    programs").

Fixes: d55a7f759dd1 ("core: probe: kernel: set the symbol address in the BPF cookie").
Signed-off-by: Antoine Tenart <atenart@redhat.com>