collect: explicit the automatic collectors selection
Auto loading and not failing on missing prerequisites for collectors was relying
on the --collectors option not being set. This is an issue when using profiles
and especially when combining them as reported by Adrián.

Making such logic explicit allows to rely on the behavior in profiles. It has
the nice consequence of also allowing to combine required and optional
collectors, e.g. "--collectors auto,ovs" where all collectors can fail to load
but the ovs one.

Reported-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: Antoine Tenart <atenart@redhat.com>