display: stop processing all the events on broken pipes
When displaying events and a broken pipe is encountered, all the remaining
events were still processed. Avoid doing so by moving the broken pipe handling
on level up.

It also has the advantage of passing down the borken pipe error in paths were we
do want to error on it (e.g. collect).

Reported-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: Antoine Tenart <atenart@redhat.com>