bindings: Update all the bindings honouring the new hooks naming scheme
Since ae9dc027e546 ("nft/ct/ovs: Add _hooks suffix to bpf hook files") hooks
got renamed, but bindings were not updated. This is not a problem, nor an
issue, but it's better keep them aligned for consistency.

Also, a weird case with mismatching bindgen version is present (e.g. dev and netns)
that uses bindgen 0.71.1 vs 0.70.1.
Given the above, and while at it, update bindgen to the latest release 0.72.1.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>