ci: add documentation checks
The test uses the same version and plugins used by the actual server.
This should make sure we have a realiable feedback in case of issues.

The container image doesn't allow us to use directly pip, but forces
the venv usage, which is not a bad thing. Let's just use that.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>