retis_in_container: fix mounting fs when using Docker and AppArmor
When using Docker as the runtime on an AppArmor enabled system, if
--allow-system-changes is enabled Retis can mount tracefs/debugfs. This will
fail with EACCES as the default Docker AppArmor profile does not allow mount
operations (as we run without --privileged).

Fix this by disabling the AppArmor profile when using Docker as the runtime,
limiting the behavior to when --allow-system-changes is enabled.

Fixes: a6dba52316d4 ("retis_in_container: be less privileged")
Signed-off-by: Antoine Tenart <atenart@redhat.com>