python: shell: print traceback on error
The Python error contains a traceback with more details than the error being
converted to anyhow. It's not possible to convey the information down the path
as it's tied to the GIL, but we can at least print it. The logic is not perfect
as PyErr helpers write directly to stderr...

Signed-off-by: Antoine Tenart <atenart@redhat.com>