compat: only apply subsequent fixups for a given version
This fixes an off by one about the fixups that are collected starting from a
given boundary. They normally should be collected for all the subsequent
versions, otherwise trying to apply fixups for existing fields may lead to
duplicate fields errors etc.

While at it, replace unneeded match/if/fallback with if/else.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>