Makefile: set version variables on demand
The version and name variables are only needed when building the Rust side of
Retis, to generate a binary. This is true for the debug and release targets, as
well as the bench one where seeing the exact version can be handy for comparing
results. Other targets do not need those variables to be set.

It's worth noting both those variables have a prerequisite on external tools
being available (rustc, cargo, localversion) and in turn computing their values
takes extra time.

Setting the two variables on demand has three advantages:
- Stop depending on external tools to be available for all other targets.
- Speed up all other targets (`make help` too).
- Speed up target completion.

Take the opportunity to expand those two variables at define time.

Signed-off-by: Antoine Tenart <atenart@redhat.com>