sock: add type and protocol
socket types have not changed in 20 years so I think it's safe to
hardcode them. Socket protocols do change, but their values are
encoded as an anonymous enum.

Therefore, using recently introduced parse_anon_enum_with to get the
names of the protocol values.

FIXME: This is terribly inefficient!! it takes 7s con the debug build.
For it to be more performant we would need to move to btf-collection and
use resolve_by_ids directly in retis code (i.e: get rid of BtfInfo
wrapper).

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>