Makefile: remove --no-gc-sections workaround
With 69c4e79c9103 ("Makefile: Instruct the linker to not gc the sections for
the test target") the linker was instructed to skip the garbage collection
for unused sections.
This was due to a problem about how .stapsdt.base section was defined and the
way most linkers acted on sections while trying to remove the unused ones.

Since 0.5.2 the probe crate no longer has the aformentioned issue.
The patch essentially reverts the workaround and updates probe to 0.5.2.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>