tools: wireshark: silence pkg-config errors
If the dependencies are not installed, all targets will print an error as the
pkg-config commands are ran while setting up variables. This is an issue when
executing the clean target, as it's included in the main clean one. We check the
availability of the dependencies in the check target already, which are only
required when building the wireshark plugin. We can thus silence the errors
while setting up variables.

Signed-off-by: Antoine Tenart <atenart@redhat.com>