collect: add command line to the startup event
While analyzing the captures it is useful to know the details of
how the tool was invoked.
This is especially true when analyzing old captures, but it is
also useful to double check that a capture was run exactly
as intended.

The patch simply adds the command line, including profiles expansion, in the
initial event.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>